Fix handling of "crashing_thread" value
This fixes the handling of the crashing_thread value so it's more like
what Socorro does. The part that was broken was that siggen defaulted
to 0 if there was no crashing_thread--it should default to None since
None indicates no crashing thread was specified and that's an important
fact in signature generation.

This also removes use of logging. We weren't really using it before.
Now we're really not using it.