Change versioning; fix dev reqs
This moves dev requirements to requirements-dev.txt. It's a little
easier to maintain them there.

This changes the versioning schema to x.y.YYYYMMDD which makes it easier
to infer the last time this was updated from the version number. That's
really helpful when you're trying to keep the lists updated.