Fix signify to convert processed crash and improve README (#91)
Previously, signify expected signature generation crash data format and
if you passed in a processed crash from Crash Stats (aka Socorro), then
it would return an empty result.

This fixes signify to recognize that situation and convert the processed
crash to a signature generation crash data format and then generate the
signature.

This also improves documentation on commands in the README.